+++ src/app/dashboard/message/components/commentItem.tsx
@@ -1,3 +1,4 @@
+import { formatDate } from "@/lib/utils";
 import { Comment } from "../interface";
 
 // CommentItem 组件的 Props 类型
@@ -10,15 +11,16 @@ interface CommentItemProps {
 
 const CommentItem = ({ comment,currentUserId,onDelete }: CommentItemProps) => {
     const isCurrentUser = currentUserId === comment.userId;
+    const formattedDate = formatDate(new Date(comment.updateTime));
     return (
-        <div className="flex items-start mb-3">
+        <div className="flex items-start mb-1">
             <img
                 src={comment.userAvatar}
                 alt={`${comment.username}'s avatar`}
                 className="w-8 h-8 rounded-full mr-3"
             />
             <div>
-                <h5 className="font-semibold text-sm">{comment.username}</h5>
+                <div className="font-semibold text-sm flex flex-row items-center">{comment.username} <div className="ml-2 text-xs text-gray-400">{formattedDate}</div></div>
                 <p className="text-gray-600 text-sm">{comment.content}</p>
             </div>
             {isCurrentUser && (
