+++ src/lib/utils.ts
@@ -25,3 +25,14 @@ export const connectToDb = async (): Promise<void> => {
     throw new Error("Could not connect to database");
   }
 };
+
+export const formatDate = (date: Date) => {
+  const year = date.getFullYear();
+  const month = String(date.getMonth() + 1).padStart(2, "0"); // Get month (0-11) and add 1
+  const day = String(date.getDate()).padStart(2, "0"); // Get day (1-31)
+  const hours = String(date.getHours()).padStart(2, "0"); // Get hours (0-23)
+  const minutes = String(date.getMinutes()).padStart(2, "0"); // Get minutes (0-59)
+  const seconds = String(date.getSeconds()).padStart(2, "0"); // Get seconds (0-59)
+
+  return `${year}-${month}-${day} ${hours}:${minutes}:${seconds}`;
+};
